requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
mysql-connector-python==9.0.0
